requirements.txt
django-crontab==0.7.1
gql[requests]==3.4.1
requests-toolbelt==1.0.0
celery==5.3.4
django-celery-beat==2.5.0
redis==5.0.1
